+++ pkg/ingester/ingester_test.go
@@ -27,7 +27,6 @@ import (
 	"github.com/weaveworks/common/user"
 	"google.golang.org/grpc"
 
-	"github.com/grafana/mimir/pkg/chunk"
 	"github.com/grafana/mimir/pkg/ingester/client"
 	"github.com/grafana/mimir/pkg/mimirpb"
 	"github.com/grafana/mimir/pkg/util/chunkcompat"
@@ -537,8 +536,29 @@ func TestIngesterMetricLimitExceeded(t *testing.T) {
 
 // Construct a set of realistic-looking samples, all with slightly different label sets
 func benchmarkData(nSeries int) (allLabels []labels.Labels, allSamples []mimirpb.Sample) {
+	// Real example from Kubernetes' embedded cAdvisor metrics, lightly obfuscated.
+	var benchmarkLabels = labels.Labels{
+		{Name: model.MetricNameLabel, Value: "container_cpu_usage_seconds_total"},
+		{Name: "beta_kubernetes_io_arch", Value: "amd64"},
+		{Name: "beta_kubernetes_io_instance_type", Value: "c3.somesize"},
+		{Name: "beta_kubernetes_io_os", Value: "linux"},
+		{Name: "container_name", Value: "some-name"},
+		{Name: "cpu", Value: "cpu01"},
+		{Name: "failure_domain_beta_kubernetes_io_region", Value: "somewhere-1"},
+		{Name: "failure_domain_beta_kubernetes_io_zone", Value: "somewhere-1b"},
+		{Name: "id", Value: "/kubepods/burstable/pod6e91c467-e4c5-11e7-ace3-0a97ed59c75e/a3c8498918bd6866349fed5a6f8c643b77c91836427fb6327913276ebc6bde28"},
+		{Name: "image", Value: "registry/organisation/name@sha256:dca3d877a80008b45d71d7edc4fd2e44c0c8c8e7102ba5cbabec63a374d1d506"},
+		{Name: "instance", Value: "ip-111-11-1-11.ec2.internal"},
+		{Name: "job", Value: "kubernetes-cadvisor"},
+		{Name: "kubernetes_io_hostname", Value: "ip-111-11-1-11"},
+		{Name: "monitor", Value: "prod"},
+		{Name: "name", Value: "k8s_some-name_some-other-name-5j8s8_kube-system_6e91c467-e4c5-11e7-ace3-0a97ed59c75e_0"},
+		{Name: "namespace", Value: "kube-system"},
+		{Name: "pod_name", Value: "some-other-name-5j8s8"},
+	}
+
 	for j := 0; j < nSeries; j++ {
-		labels := chunk.BenchmarkLabels.Copy()
+		labels := benchmarkLabels.Copy()
 		for i := range labels {
 			if labels[i].Name == "cpu" {
 				labels[i].Value = fmt.Sprintf("cpu%02d", j)